image-processor-aws/lambda_function/requirements.txt
Pillow>=9.0
boto3>=1.20